PianoRisePython/requirements.txt
flask
flask-Login
flask-socketio
flask-sqlalchemy
flask-migrate
flask-wtf
werkzeug
psycopg2-binary
numpy
scipy
librosa
pylint
pyan3==1.1.1
graphviz
pydeps
python-dotenv
